requirements.txt
aiohttp==3.9.5
beautifulsoup4==4.12.3
discord.py==2.3.2
emoji==2.11.1
peewee==3.17.3
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.31.0

aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pycparser==2.22
soupsieve==2.5
urllib3==2.2.1
yarl==1.9.4
